backend/requirements.txt
sqlalchemy
fastapi
uvicorn
python-multipart
pydantic
pyjwt
python-jose
passlib
bcrypt
